requirements.txt
fastapi
uvicorn[standard]
gunicorn
psycopg2-binary
celery
pydantic-settings
redis>=4.0
alembic
pgvector
asyncpg 
sqlalchemy[asyncio]
qdrant-client[grpc,aio]==1.9.0
google-generativeai
structlog
starlette-prometheus
passlib[bcrypt]
python-jose
Authlib
pydantic[email]
python-multipart
starlette
itsdangerous
reportlab
requests
mcp>=1.0.0
groq>=0.4.0
playwright>=1.40.0
python-dotenv>=1.0.0